src/requirements.txt
Telethon==1.27.0
python-dotenv==1.0.0
# requests-toolbelt==0.10.1
requests==2.28.2
Flask==2.2.3
gunicorn==20.1.0
